dd data type descriptions
AMD matrix instructions include input and output type specifiers
in the instruction mnemonic, but these are not always the clearest
way to describe what data types as used for input and output.

Add a new detailed instruction print section that describes the
data type used by the source and destination registers.