Add support for MI300
Adds support for the matrix core instructions in the MI300 series of
accelerators. These accelerators, sometimes referred to in the kernel
mode driver as aqua vanjaram, use the CDNA3 instruction set, also
known as gfx940, gfx941, and gfx942 in the compiler.

This includes different throughputs and MxNxK sizes for existing data
types, as well as support for:
 - XF32 (i.e., FP32 input/output with lower-precision internal math)
 - 8-bit floating point formats (FP8 and BF8)
 - 4:2 structured sparsity (SMFMAC operations)
 - Negating FP64 inputs using the BLGP modifier