Add support for RDNA4 / gfx12
Adds support for the AI accelerator instructions in the RDNA4
architecture GPUs. At the moment, these GPUs include the specific
ISAs gfx1200 and gfx1201. The AI accelerator instructions in both
of these specific ISAs are the same, so they are treated the same
in this matrix calculator tool.

Compared to the AI accelerator instructions in RDNA3, RDNA4 includes:
 - A new register storage layout, including the ability to
   use all lanes to hold different data.
 - Higher computational throughput rates
 - OCP 8-bit floating point formats (FP8 and BF8)
 - Support for 4:2 structured sparsity (SWMMAC operations)
 - New 4-bit integer instruction with 2x higher K dimension