Fix CBSZ for CDNA3 sparse compression indices
Previous versions of the matrix instruction calculator incorrectly
stated that, in CDNA3, when attempting to use the ABID modifier to
change which bits of the Src2 ArchVGPR were used for the sparse
matrix compression index values, the CBSZ modifier must be non-zero.
This contradicted the public CDNA3 ISA guide, which states that
the ABID values are only followed with CBSZ=0. When CBSZ!=0, the
ABID field is treated as if it were zero for the compression index
offsetting.

The ISA guide matches actual hardware capabilities, and the matrix
instruction calculator (and README file) were incorrect.
This patch updates the README file, the matrix instruction calculator,
and the delta tester to properly handle this situtation